refactor: share terminal options via node-free module
